Split mixer.c into mixer.c and servos.c
