+++ src/main/blackbox/blackbox.c
@@ -54,6 +54,7 @@
 
 
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/altitudehold.h"
 #include "flight/failsafe.h"
 #include "flight/imu.h"
