+++ src/main/build_config.c
@@ -26,6 +26,7 @@
 #include "drivers/timer.h"
 #include "drivers/pwm_mapping.h"
 #include "flight/mixer.h"
+#include "flight/servos.h"
 
 #include "build_config.h"
