+++ src/main/flight/mixer.h
@@ -19,15 +19,12 @@
 
 #if defined(USE_QUAD_MIXER_ONLY)
 #define MAX_SUPPORTED_MOTORS 4
-#define MAX_SUPPORTED_SERVOS 1
 
 #elif defined(TARGET_MOTOR_COUNT)
 #define MAX_SUPPORTED_MOTORS TARGET_MOTOR_COUNT
-#define MAX_SUPPORTED_SERVOS 8
 
 #else
 #define MAX_SUPPORTED_MOTORS 12
-#define MAX_SUPPORTED_SERVOS 8
 #endif
 
 #define YAW_JUMP_PREVENTION_LIMIT_LOW 80
@@ -104,113 +101,6 @@ PG_DECLARE(motor3DConfig_t, motor3DConfig);
 
 #define CHANNEL_FORWARDING_DISABLED (uint8_t)0xFF
 
-#ifdef USE_SERVOS
-
-// These must be consecutive, see 'reversedSources'
-enum {
-    INPUT_STABILIZED_ROLL = 0,
-    INPUT_STABILIZED_PITCH,
-    INPUT_STABILIZED_YAW,
-    INPUT_STABILIZED_THROTTLE,
-    INPUT_RC_ROLL,
-    INPUT_RC_PITCH,
-    INPUT_RC_YAW,
-    INPUT_RC_THROTTLE,
-    INPUT_RC_AUX1,
-    INPUT_RC_AUX2,
-    INPUT_RC_AUX3,
-    INPUT_RC_AUX4,
-    INPUT_GIMBAL_PITCH,
-    INPUT_GIMBAL_ROLL,
-
-    INPUT_SOURCE_COUNT
-} inputSource_e;
-
-// target servo channels
-typedef enum {
-    SERVO_GIMBAL_PITCH = 0,
-    SERVO_GIMBAL_ROLL = 1,
-    SERVO_ELEVATOR = 2,
-    SERVO_FLAPPERON_1 = 3,
-    SERVO_FLAPPERON_2 = 4,
-    SERVO_RUDDER = 5,
-    SERVO_THROTTLE = 6, // for internal combustion (IC) planes
-    SERVO_FLAPS = 7,
-
-    SERVO_BICOPTER_LEFT = 4,
-    SERVO_BICOPTER_RIGHT = 5,
-
-    SERVO_DUALCOPTER_LEFT = 4,
-    SERVO_DUALCOPTER_RIGHT = 5,
-
-    SERVO_SINGLECOPTER_1 = 3,
-    SERVO_SINGLECOPTER_2 = 4,
-    SERVO_SINGLECOPTER_3 = 5,
-    SERVO_SINGLECOPTER_4 = 6,
-
-} servoIndex_e; // FIXME rename to servoChannel_e
-
-#define SERVO_PLANE_INDEX_MIN SERVO_ELEVATOR
-#define SERVO_PLANE_INDEX_MAX SERVO_FLAPS
-
-#define SERVO_DUALCOPTER_INDEX_MIN SERVO_DUALCOPTER_LEFT
-#define SERVO_DUALCOPTER_INDEX_MAX SERVO_DUALCOPTER_RIGHT
-
-#define SERVO_SINGLECOPTER_INDEX_MIN SERVO_SINGLECOPTER_1
-#define SERVO_SINGLECOPTER_INDEX_MAX SERVO_SINGLECOPTER_4
-
-#define SERVO_FLAPPERONS_MIN SERVO_FLAPPERON_1
-#define SERVO_FLAPPERONS_MAX SERVO_FLAPPERON_2
-
-typedef struct servoMixer_s {
-    uint8_t targetChannel;                  // servo that receives the output of the rule
-    uint8_t inputSource;                    // input channel for this rule
-    int8_t rate;                            // range [-125;+125] ; can be used to adjust a rate 0-125% and a direction
-    uint8_t speed;                          // reduces the speed of the rule, 0=unlimited speed
-    int8_t min;                             // lower bound of rule range [0;100]% of servo max-min
-    int8_t max;                             // lower bound of rule range [0;100]% of servo max-min
-    uint8_t box;                            // active rule if box is enabled, range [0;3], 0=no box, 1=BOXSERVO1, 2=BOXSERVO2, 3=BOXSERVO3
-} servoMixer_t;
-
-#define MAX_SERVO_RULES (2 * MAX_SUPPORTED_SERVOS)
-#define MAX_SERVO_SPEED UINT8_MAX
-#define MAX_SERVO_BOXES 3
-
-PG_DECLARE_ARR(servoMixer_t, MAX_SERVO_RULES, customServoMixer);
-
-// Custom mixer configuration
-typedef struct mixerRules_s {
-    uint8_t servoRuleCount;
-    const servoMixer_t *rule;
-} mixerRules_t;
-
-typedef struct servoParam_s {
-    int16_t min;                            // servo min
-    int16_t max;                            // servo max
-    int16_t middle;                         // servo middle
-    int8_t rate;                            // range [-125;+125] ; can be used to adjust a rate 0-125% and a direction
-    uint8_t angleAtMin;                     // range [0;180] the measured angle in degrees from the middle when the servo is at the 'min' value.
-    uint8_t angleAtMax;                     // range [0;180] the measured angle in degrees from the middle when the servo is at the 'max' value.
-    int8_t forwardFromChannel;              // RX channel index, 0 based.  See CHANNEL_FORWARDING_DISABLED
-    uint32_t reversedSources;               // the direction of servo movement for each input source of the servo mixer, bit set=inverted
-} __attribute__ ((__packed__)) servoParam_t;
-
-typedef struct servoProfile_s {
-    servoParam_t servoConf[MAX_SUPPORTED_SERVOS];
-} servoProfile_t;
-
-PG_DECLARE_PROFILE(servoProfile_t, servoProfile);
-
-struct gimbalConfig_s;
-struct motorAndServoConfig_s;
-struct rxConfig_s;
-
-extern int16_t servo[MAX_SUPPORTED_SERVOS];
-bool isMixerUsingServos(void);
-void writeServos(void);
-void filterServos(void);
-#endif
-
 extern int16_t motor[MAX_SUPPORTED_MOTORS];
 extern int16_t motor_disarmed[MAX_SUPPORTED_MOTORS];
 
@@ -219,15 +109,9 @@ extern bool motorLimitReached;
 void mixerInit(motorMixer_t *customMotorMixers);
 void writeAllMotors(int16_t mc);
 void mixerLoadMix(int index, motorMixer_t *customMixers);
-#ifdef USE_SERVOS
-void mixerInitServos(servoMixer_t *customServoMixers);
-void mixerUseConfigs(servoParam_t *servoConfToUse);
-void servoMixerLoadMix(int index, servoMixer_t *customServoMixers);
-void loadCustomServoMixer(void);
-int servoDirection(int servoIndex, int fromChannel);
-#endif
 void mixerResetDisarmedMotors(void);
 void mixTable(void);
+void servoMixTable(void);
 void writeMotors(void);
 void stopMotors(void);
 void StopPwmAllMotors(void);
