+++ src/main/io/msp.c
@@ -80,6 +80,7 @@
 #include "sensors/gyro.h"
 
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/failsafe.h"
