+++ src/test/unit/config_eeprom_unittest.cc
@@ -58,6 +58,7 @@ extern "C" {
     #include "sensors/boardalignment.h"
 
     #include "flight/mixer.h"
+    #include "flight/servos.h"
     #include "flight/imu.h"
     #include "flight/navigation.h"
     #include "flight/failsafe.h"
