+++ src/test/unit/flight_mixer_unittest.cc
@@ -45,6 +45,7 @@ extern "C" {
     #include "flight/pid.h"
     #include "flight/imu.h"
     #include "flight/mixer.h"
+    #include "flight/servos.h"
 
     #include "io/motor_and_servo.h"
     #include "io/gimbal.h"
