+++ src/test/unit/msp_unittest.cc
@@ -68,6 +68,7 @@ extern "C" {
     #include "sensors/compass.h"
 
     #include "flight/mixer.h"
+    #include "flight/servos.h"
     #include "flight/pid.h"
     #include "flight/navigation.h"
     #include "flight/imu.h"
